create local storage util

+++ client/src/app/utils/local-storage.util.ts
@@ -1,44 +1,56 @@
-// src/app/utils/local-storage.util.ts
-
 export class LocalStorageUtil {
+	private static isBrowser(): boolean {
+		return typeof window !== 'undefined';
+	}
+
 	public static Supported(): boolean {
-		return typeof Storage !== 'undefined';
+		return LocalStorageUtil.isBrowser() && typeof Storage !== 'undefined';
 	}
 
 	public static set(key: string, value: any, ttl = 0): void {
-		const item = {
-			value,
-			canExpire: ttl >= 0,
-			expireAt: (new Date().getTime()) / 1000 + ttl,
-		};
+		if (LocalStorageUtil.isBrowser()) {
+			const item = {
+				value,
+				canExpire: ttl >= 0,
+				expireAt: (new Date().getTime()) / 1000 + ttl,
+			};
 
-		localStorage.setItem(key, JSON.stringify(item));
+			localStorage.setItem(key, JSON.stringify(item));
+		}
 	}
 
 	public static get(key: string): any {
-		const itemStr = localStorage.getItem(key);
+		if (LocalStorageUtil.isBrowser()) {
+			const itemStr = localStorage.getItem(key);
 
-		if (!itemStr) {
-			return null;
+			if (!itemStr) {
+				return null;
+			}
+
+			const item = JSON.parse(itemStr);
+
+			if (item.canExpire && (new Date().getTime()) / 1000 > item.expireAt) {
+				localStorage.removeItem(key);
+				return null;
+			}
+
+			return item.value;
 		}
 
-		const item = JSON.parse(itemStr);
-
-		if (item.canExpire && (new Date().getTime()) / 1000 > item.expireAt) {
-			localStorage.removeItem(key);
-			return null;
-		}
-
-		return item.value;
+		return null;
 	}
 
 	public static clearAll(): void {
-		localStorage.clear();
-		console.log('LocalStorageUtil: Localstorage has been cleared');
+		if (LocalStorageUtil.isBrowser()) {
+			localStorage.clear();
+			console.log('LocalStorageUtil: Localstorage has been cleared');
+		}
 	}
 
 	public static clear(key: string): void {
-		localStorage.removeItem(key);
-		console.log(`LocalStorageUtil: ${key} has been removed from localstorage`);
+		if (LocalStorageUtil.isBrowser()) {
+			localStorage.removeItem(key);
+			console.log(`LocalStorageUtil: ${key} has been removed from localstorage`);
+		}
 	}
 }
